removing software is small, data is large statement
while I still think there is something to this argument, I don't know how to state it well, and it a generalization that is not true in many specific cases, so it seems best to remove it at this point.  If someone can define something like this more concisely and more accurately, it can always come back.